Added ability to fill in time series
I added code to fill in any missing dates in our timeseries. Currently problamatic because the minutes and seconds are variable in the data which makes it hard to prevent duplicates. Will consult on if we want to round up data to hours. Also altered date filtering to wokr in Posixlt. May move later step that converts to Posixct to precede this step