Fixing Timezone/DST issue
The metstations are ignoring daylight savings time.  So we're forcing it to be ignored in the data by setting timezone to UTC+6.  Per R documentation, positive offsets are behind (WEST of) UTC: https://stat.ethz.ch/R-manual/R-devel/library/base/html/timezones.html

(Such a pain in the ass... why is DST still a thing???)